weathersite/requirements.txt
-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
django-ipware==3.0.2
django==3.2
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.4
pyjwt==2.0.1; python_version >= '3.6'
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
